[#2126] - Extrae la presentación del reporte a su propio módulo
El núcleo de clasificación quedó rozando el techo de 500 líneas, y lo que
sobraba respondía otra pregunta: cómo se lee lo medido, no qué significa. La
separación es la misma que ya tiene la política de reintento, y deja lugar
para una segunda forma de presentar la misma corrida.

Movimiento puro: ninguna función cambia, y sus tests se mudan con ella.
